+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 7.1.4 January 5, 2026
+
+Added `denom` argument to `pct` in utils.py to allow for not including
+the denominator in the result.
+
 ## 7.1.3 January 5, 2026
 
 Added `--listReferences` option to `sam-coverage.py` to just list the